@@ -1,46 +1,62 @@
 {
-  "name": "frontend",
+  "name": "insurance-core",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --proxy-config proxy.conf.json",
+    "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.8",
-    "@angular/cdk": "^17.3.9",
-    "@angular/common": "^17.3.8",
-    "@angular/compiler": "^17.3.8",
-    "@angular/core": "^17.3.8",
-    "@angular/forms": "^17.3.8",
-    "@angular/material": "^17.3.9",
-    "@angular/material-luxon-adapter": "^17.3.9",
-    "@angular/platform-browser": "^17.3.8",
-    "@angular/platform-browser-dynamic": "^17.3.8",
-    "@angular/router": "^17.3.8",
-    "@nestjs/apollo": "^12.1.0",
-    "apollo-server-express": "^3.13.0",
-    "graphql": "^16.8.1",
+    "@angular/animations": "^18.0.0",
+    "@angular/cdk": "^18.0.0",
+    "@angular/common": "^18.0.0",
+    "@angular/compiler": "^18.0.0",
+    "@angular/core": "^18.0.0",
+    "@angular/forms": "^18.0.0",
+    "@angular/localize": "^18.0.0",
+    "@angular/material": "^18.0.0",
+    "@angular/material-luxon-adapter": "^18.0.0",
+    "@angular/platform-browser": "^18.0.0",
+    "@angular/platform-browser-dynamic": "^18.0.0",
+    "@angular/router": "^18.0.0",
+    "@types/quill": "^2.0.14",
     "luxon": "^3.4.4",
+    "ngx-cookie-service": "^18.0.0",
+    "ngx-quicklink": "^0.4.5",
+    "ngx-quill": "^25.3.2",
+    "ngx-toastr": "^18.0.0",
+    "quill": "^2.0.2",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.5"
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.7",
-    "@angular/cli": "~17.3.7",
-    "@angular/compiler-cli": "^17.3.8",
-    "@types/jasmine": "~4.3.0",
+    "@angular-devkit/build-angular": "^18.0.0",
+    "@angular-eslint/builder": "17.5.0",
+    "@angular-eslint/eslint-plugin": "17.5.0",
+    "@angular-eslint/eslint-plugin-template": "17.5.0",
+    "@angular-eslint/schematics": "17.5.0",
+    "@angular-eslint/template-parser": "17.5.0",
+    "@angular/cli": "^18.0.0",
+    "@angular/compiler-cli": "^18.0.0",
+    "@types/jasmine": "~5.1.0",
     "@types/luxon": "^3.4.2",
-    "jasmine-core": "~4.5.0",
+    "@typescript-eslint/eslint-plugin": "7.8.0",
+    "@typescript-eslint/parser": "7.8.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "jasmine-core": "~5.1.0",
     "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~5.4.5"
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "prettier": "^3.2.5",
+    "prettier-eslint": "^16.3.0",
+    "typescript": "~5.4.2"
   }
 }
