@@ -1,39 +1,42 @@
-// app.component.ts
-import { Component, OnInit } from "@angular/core";
-import { ThemeService } from './services/theme.service';
-import { Router, RouterOutlet } from "@angular/router";
-import { AuthService } from "./services/auth.service";
-import { NavigationComponent } from "./navigation/navigation.component";
-import { CategoryService } from "./services/category.service";
+import { Component, Renderer2, effect, inject } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+import { DOCUMENT } from '@angular/common';
+import { ThemeService } from './core/services/theme.service';
+import { LayoutComponent } from './main/layout/layout.component';
+import { MatToolbar } from '@angular/material/toolbar';
+import { MatButtonToggle, MatButtonToggleGroup } from '@angular/material/button-toggle';
+import { MatFormField, MatLabel } from '@angular/material/form-field';
+import { MatDatepicker, MatDatepickerInput, MatDatepickerToggle } from '@angular/material/datepicker';
+import { provideLuxonDateAdapter } from '@angular/material-luxon-adapter';
+import { MatInput } from '@angular/material/input';
 
 @Component({
-    selector: 'app-root',
-    templateUrl: './app.component.html',
-    standalone: true,
-  imports: [RouterOutlet, NavigationComponent]
+  selector: 'app-root',
+  standalone: true,
+  imports: [
+    RouterOutlet,
+    LayoutComponent,
+    MatToolbar,
+    MatButtonToggleGroup,
+    MatButtonToggle,
+    MatFormField,
+    MatLabel,
+    MatDatepickerInput,
+    MatDatepickerToggle,
+    MatDatepicker,
+    MatInput
+  ],
+  providers: [provideLuxonDateAdapter()],
+  templateUrl: './app.component.html'
 })
-export class AppComponent implements OnInit {
-  constructor(
-    private themeService: ThemeService,
-    private router: Router,
-    private authService: AuthService,
-    private categoryService: CategoryService
-  ) {}
+export class AppComponent {
+  #themeService: ThemeService = inject(ThemeService);
+  #document: Document = inject(DOCUMENT);
+  #renderer: Renderer2 = inject(Renderer2);
 
-  ngOnInit(): void {
-    this.themeService.initializeTheme(document.body);
-    this.categoryService.getAllCategories().subscribe(res => {
-      console.log(res);
-    });
-  }
-
-  private checkAuthentication(): void {
-    this.authService.isAuthenticated().subscribe(isAuthenticated => {
-      if (isAuthenticated) {
-        this.router.navigate(["/dashboard"]);
-      } else {
-        this.router.navigate(["/sign-in"]);
-      }
+  constructor() {
+    effect((): void => {
+      this.#renderer.setAttribute(this.#document.documentElement, 'class', this.#themeService.theme());
     });
   }
 }
