@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { AuthService } from "../../services/auth.service";
+import { AuthService } from "../../core/services/auth.service";
 import { MatSnackBar, MatSnackBarRef, TextOnlySnackBar } from "@angular/material/snack-bar";
 import { Router } from "@angular/router";
 import { AuthComponent } from '../auth.component';
